Separate out QueueList
